+++ src/Exception/Logical/InvalidControllerException.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types = 1);
+
+namespace Apitte\Core\Exception\Logical;
+
+use Apitte\Core\Exception\LogicalException;
+
+class InvalidControllerException extends LogicalException
+{
+
+}
